+++ nrf_802154/driver/include/nrf_802154_types.h
@@ -36,6 +36,7 @@
 #define NRF_802154_TYPES_H__
 
 #include <stdint.h>
+#include <nrf_802154_config.h>
 
 /**
  * Avoid including nrfx dependencies in a unit test build.
@@ -473,6 +474,17 @@ typedef void (* nrf_802154_transmit_failed_notification_t)(
     uint8_t                                   * p_frame,
     nrf_802154_tx_error_t                       error,
     const nrf_802154_transmit_done_metadata_t * p_meta);
+
+#if (NRF_802154_ENERGY_DETECTED_VERSION != 0)
+/**
+ * @brief Structure that holds results of energy detection procedure.
+ */
+typedef struct
+{
+    int8_t ed_dbm; // !< Maximum detected ED in dBm.
+} nrf_802154_energy_detected_t;
+
+#endif
 
 /**
  *@}