+++ nrf_802154/driver/src/nrf_802154.c
@@ -201,10 +201,20 @@ void nrf_802154_short_address_set(const uint8_t * p_short_address)
     nrf_802154_pib_short_address_set(p_short_address);
 }
 
+#if (NRF_802154_ENERGY_DETECTED_VERSION != 0)
+uint8_t nrf_802154_energy_level_from_dbm_calculate(int8_t ed_dbm)
+{
+    return nrf_802154_addons_energy_level_from_dbm_calculate(ed_dbm);
+}
+
+#else
+
 int8_t nrf_802154_dbm_from_energy_level_calculate(uint8_t energy_level)
 {
     return nrf_802154_addons_dbm_from_energy_level_calculate(energy_level);
 }
+
+#endif // NRF_802154_ENERGY_DETECTED_VERSION != 0
 
 uint8_t nrf_802154_ccaedthres_from_dbm_calculate(int8_t dbm)
 {
@@ -1217,10 +1227,19 @@ __WEAK void nrf_802154_transmit_failed(uint8_t                                  
     (void)p_metadata;
 }
 
+#if (NRF_802154_ENERGY_DETECTED_VERSION != 0)
+__WEAK void nrf_802154_energy_detected(const nrf_802154_energy_detected_t * p_result)
+{
+    (void)p_result;
+}
+
+#else
 __WEAK void nrf_802154_energy_detected(uint8_t result)
 {
     (void)result;
 }
+
+#endif
 
 __WEAK void nrf_802154_energy_detection_failed(nrf_802154_ed_error_t error)
 {