+++ nrf_802154/serialization/include/host/nrf_802154.h
@@ -703,6 +703,18 @@ void nrf_802154_tx_power_set(int8_t power);
  */
 int8_t nrf_802154_tx_power_get(void);
 
+#if (NRF_802154_ENERGY_DETECTED_VERSION != 0)
+/**
+ * @brief  Converts energy level received during the energy detection procedure to IEEE Std. 802.15.4-2015 compliant value.
+ *
+ * @param ed_dbm  Energy level in dBm
+ *
+ * @return uint8_t  Energy level in units compliant to IEEE Std. 802.15.4-2015 chapter 10.2.5.
+ */
+uint8_t nrf_802154_energy_level_from_dbm_calculate(int8_t ed_dbm);
+
+#else
+
 /**
  * @brief  Converts the energy level received during the energy detection procedure to a dBm value.
  *
@@ -711,6 +723,8 @@ int8_t nrf_802154_tx_power_get(void);
  * @return  Result of the energy detection procedure in dBm.
  */
 int8_t nrf_802154_dbm_from_energy_level_calculate(uint8_t energy_level);
+
+#endif // NRF_802154_ENERGY_DETECTED_VERSION != 0
 
 /**
  * @brief  Calculates the timestamp of the first symbol of the preamble in a received frame.